fix: reverse proxy unescape the URI
When there are escape sequences in the URL they were unescaped before forwarding the request to the
internal services. This is especially problematic when `%2F` is used which is turned into a `/`.

There doesn't seem to be a solution to this problem in nginx. It can forward the URI unmodified,
but as soon as you try to the prefix all options will unescape the URI.

It's hard to understand why nginx is so popular, but at the same time it's so difficult to configure
it for these common use cases.

Just compare the previous nginx config and the new caddy config. It's so much more straight forward
and just works. Even websockets don't need any special configuration.