feat!: merged the modelix.workspaces repository
https://github.com/modelix/modelix.workspaces/commit/fb0476638b3187c461405a24b8815f6ffd45febc

The workspaces project is tightly coupled with the helm chart so that a modelix.workspaces release
almost always requires a release in modelix.kubernetes. The workspaces are the main feature of the
helm chart so that there rarely is a release of modelix.kubernetes without a change in
modelix.workspaces.

Merging the two repositories makes releases easier and integration test can already run for a PR
that changes something in the workspaces.

Marked as breaking change to bump the version to 1.0.0, which is newer for both repositories.
The last release in modelix.workspaces is 0.15.0 and in modelix.kubernetes 0.12.0.